api/requirements.txt
flask
numpy
pandas
python-dotenv
sodapy
google-cloud-storage